+++ lib/components/JSONTextFormField.dart
@@ -1,8 +1,13 @@
+import 'dart:io';
+
 import 'package:barcode_scan/barcode_scan.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:json_schema_form/models/Action.dart';
 import 'package:json_schema_form/models/Schema.dart';
+import 'package:file_chooser/file_chooser.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 class JSONTextFormField extends StatefulWidget {
   final Schema schema;
@@ -44,18 +49,28 @@ class _JSONTextFormFieldState extends State<JSONTextFormField> {
     }
   }
 
-  _suffixIconAction({dynamic image, dynamic inputValue}) {
+  _suffixIconAction({File image, String inputValue}) async {
     switch (widget.schema.action.actionDone) {
       case ActionDone.getInput:
-        setState(() {
-          _controller.text = inputValue.toString();
-        });
-
+        if (inputValue != null) {
+          setState(() {
+            _controller.text = inputValue.toString();
+          });
+        } else if (image != null) {
+          var value =
+              await (widget.schema.action as FieldAction<File>).onDone(image);
+          if (value is String) {
+            setState(() {
+              _controller.text = value;
+            });
+          }
+        }
         break;
 
       case ActionDone.getImage:
-
-        // TODO: Add image support
+        if (image != null) {
+          await (widget.schema.action as FieldAction<File>).onDone(image);
+        }
         break;
     }
   }
@@ -64,18 +79,65 @@ class _JSONTextFormFieldState extends State<JSONTextFormField> {
     if (widget.schema.action != null) {
       switch (widget.schema.action.actionTypes) {
         case ActionTypes.image:
-          break;
+          return IconButton(
+            onPressed: () {
+              showModalBottomSheet(
+                context: context,
+                builder: (ctx) => Container(
+                  child: Wrap(
+                    children: <Widget>[
+                      Platform.isAndroid || Platform.isIOS
+                          ? ListTile(
+                              leading: Icon(Icons.camera_alt),
+                              title: Text("From Camera"),
+                              onTap: () async {
+                                File file = await ImagePicker.pickImage(
+                                    source: ImageSource.camera);
+                                await _suffixIconAction(image: file);
+                              },
+                            )
+                          : null,
+                      ListTile(
+                        leading: Icon(Icons.filter),
+                        title: Text("From Gallery"),
+                        onTap: () async {
+                          if (Platform.isIOS || Platform.isAndroid) {
+                            File file = await ImagePicker.pickImage(
+                                source: ImageSource.gallery);
+                            await _suffixIconAction(image: file);
+                          } else if (Platform.isMacOS) {
+                            var result = await showOpenPanel();
+                            if (!result.canceled) {
+                              if (result.paths.length > 0) {
+                                await _suffixIconAction(
+                                    image: File(result.paths.first));
+                              }
+                            }
+                          }
+                        },
+                      )
+                    ],
+                  ),
+                ),
+              );
+            },
+            icon: Icon(Icons.camera_alt),
+          );
 
         case ActionTypes.qrScan:
           return IconButton(
             onPressed: () async {
-              try {
-                String barcode = await BarcodeScanner.scan();
-                _suffixIconAction(inputValue: barcode);
-              } on PlatformException catch (e) {
-                if (e.code == BarcodeScanner.CameraAccessDenied) {
-                } else {}
-              } on FormatException {} catch (e) {}
+              if (Platform.isAndroid || Platform.isIOS) {
+                try {
+                  String barcode = await BarcodeScanner.scan();
+                  await _suffixIconAction(inputValue: barcode);
+                } on PlatformException catch (e) {
+                  if (e.code == BarcodeScanner.CameraAccessDenied) {
+                  } else {}
+                } on FormatException {} catch (e) {}
+              } else if (Platform.isMacOS) {
+                //TODO: Add macOS support
+              }
             },
             icon: Icon(Icons.camera_alt),
           );